fix(release): keep release tool and tests within source-length budget
The generic nested-group scan pushed tools/moon/cmd/release/main.mbt past its base line count and tests/tooling/moonbit-release.test.ts across the 350-line gate, failing source-file-lengths. Split the native-binaries catalog rewriters into a sibling same-package file and move the release-guard fixture into tests/tooling/support, both behavior-preserving, to bring each file back under budget.
