+++ tests/tooling/moonbit-release.test.ts
@@ -6,7 +6,10 @@ import path from "node:path";
 import { test } from "node:test";
 
 import { repoRoot, runMoonScript } from "./_helpers/moonbit.ts";
-import { writeFakeCommand } from "./support/fake-command.ts";
+import {
+  runRepositoryGuardFixture,
+  type RepositoryGuardOptions,
+} from "./support/release-guard-fixture.ts";
 
 function writeTempFile(contents: string): string {
   const dir = fs.mkdtempSync(path.join(os.tmpdir(), "vize-release-test-"));
@@ -242,80 +245,6 @@ test("release script rejects the removed force-tag escape hatch", () => {
   assert.equal(result.status, 1);
   assert.match(result.stderr, /--force-tag is not supported; published release tags are immutable/);
 });
-
-interface RepositoryGuardOptions {
-  branch: string;
-  dirty?: boolean;
-  ancestor?: boolean;
-  headSha?: string;
-  remoteSha?: string;
-  localTagExists?: boolean;
-  remoteTagExists?: boolean;
-  pushFails?: boolean;
-  stagedFiles?: boolean;
-  manifestTestFails?: boolean;
-}
-
-function runRepositoryGuardFixture(options: RepositoryGuardOptions) {
-  const tempDir = fs.mkdtempSync(path.join(os.tmpdir(), "vize-release-guard-"));
-  const binDir = path.join(tempDir, "bin");
-  const gitLogPath = path.join(tempDir, "git.log");
-  const cargoTomlPath = path.join(tempDir, "Cargo.toml");
-  const cargoToml = '[workspace.package]\nversion = "0.290.0"\n';
-  fs.mkdirSync(binDir, { recursive: true });
-  fs.mkdirSync(path.join(tempDir, "npm"));
-  fs.mkdirSync(path.join(tempDir, "tests/tooling"), { recursive: true });
-  fs.writeFileSync(gitLogPath, "");
-  fs.writeFileSync(cargoTomlPath, cargoToml);
-  fs.writeFileSync(path.join(tempDir, "pnpm-workspace.yaml"), "");
-  fs.writeFileSync(path.join(tempDir, "pnpm-lock.yaml"), "");
-  fs.writeFileSync(
-    path.join(tempDir, "tests/tooling/package-manifests.test.ts"),
-    options.manifestTestFails ? 'throw new Error("manifest drift");\n' : "",
-  );
-  writeFakeCommand(binDir, "cargo", "process.exit(0);");
-  writeFakeCommand(
-    binDir,
-    "git",
-    [
-      "const fs = require('node:fs');",
-      "const args = process.argv.slice(2);",
-      "fs.appendFileSync(process.env.GIT_LOG, args.join(' ') + '\\n');",
-      "if (args[0] === 'branch') { console.log(process.env.TEST_BRANCH); process.exit(0); }",
-      "if (args[0] === 'status') { if (process.env.TEST_DIRTY === 'true') console.log(' M Cargo.toml'); process.exit(0); }",
-      "if (args[0] === 'fetch') process.exit(0);",
-      "if (args[0] === 'merge-base') process.exit(process.env.TEST_ANCESTOR === 'false' ? 1 : 0);",
-      "if (args[0] === 'rev-parse' && args.includes('--verify')) process.exit(process.env.LOCAL_TAG_EXISTS === 'true' ? 0 : 1);",
-      "if (args[0] === 'rev-parse') { console.log(args.at(-1) === 'HEAD' ? process.env.TEST_HEAD_SHA : process.env.TEST_REMOTE_SHA); process.exit(0); }",
-      "if (args[0] === 'ls-remote' && process.env.REMOTE_TAG_EXISTS === 'true') { console.log('bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb\\t' + args.at(-1)); process.exit(0); }",
-      "if (args[0] === 'ls-remote') process.exit(2);",
-      "if (args[0] === 'diff' && args.includes('--cached')) { if (process.env.TEST_STAGED_FILES !== 'false') console.log('Cargo.toml'); process.exit(0); }",
-      "if (args[0] === 'push') process.exit(process.env.TEST_PUSH_FAIL === 'true' ? 1 : 0);",
-      "if (['add', 'commit', 'tag'].includes(args[0])) process.exit(0);",
-      "process.exit(1);",
-    ].join("\n"),
-  );
-
-  const result = runMoonScript("release", ["patch", "-y"], {
-    cwd: tempDir,
-    env: {
-      PATH: `${binDir}${path.delimiter}${process.env.PATH ?? ""}`,
-      GIT_LOG: gitLogPath,
-      TEST_BRANCH: options.branch,
-      TEST_DIRTY: String(options.dirty ?? false),
-      TEST_ANCESTOR: String(options.ancestor ?? true),
-      TEST_HEAD_SHA: options.headSha ?? "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-      TEST_REMOTE_SHA: options.remoteSha ?? "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-      LOCAL_TAG_EXISTS: String(options.localTagExists ?? false),
-      REMOTE_TAG_EXISTS: String(options.remoteTagExists),
-      TEST_PUSH_FAIL: String(options.pushFails ?? false),
-      TEST_STAGED_FILES: String(options.stagedFiles ?? true),
-      VIZE_RELEASE_GUARD_SCRIPT: path.join(repoRoot, "tools/github/release-local-guard.mjs"),
-    },
-  });
-  const gitLog = fs.readFileSync(gitLogPath, "utf8");
-  return { cargoToml, cargoTomlPath, gitLog, result, tempDir };
-}
 
 test("release script refuses to create an empty release commit", () => {
   const fixture = runRepositoryGuardFixture({ branch: "main", stagedFiles: false });
